Allows for model/db TIME fields. Previously PHPR would treat database TIME fields as a datetime, causing errors if time fields were used in models.
This update allows for a time field value to be stored in the database as entered by the user (not treated as datetime and GMT-ZEROD).

This is useful in situations where you would like to store a local time for a future event. Currently datetime will always convert to GMT-0 for internationalisation,
but this is not suited to scheduling future times and dates in areas with daylight savings. For example if you have an event scheduled for 15:00 12/12/2014 using datetime whilst in DST+1, the time will be stored in the database as 14:00 12/12/2014. This is fine whilst you remain in daylight savings but the scheduled events time will change to 14:00 when DST expires.  With the TIME field available you have the option to schedule in local time buy using seperate DATE, TIME and TIMEZONE fields.